Module Block.Footnote

Footnotes.

Footnotes

type block := t

type t

The type for footnotes.

val make : ?indent:Layout.indent -> ?defined_label:Label.t option -> Label.t -> block -> t

make label b is a footnote for label label with content b. defined_label defaults to label.

val indent : t -> Layout.indent

indent fn is the indentation to the label found on the first line.

val label : t -> Label.t

label fn is the footnote definition label as found in the source text. It includes the ^. See also defined_label.

val defined_label : t -> Label.t option

defined_label fn is the label determined by the Label.resolver for the footnote. The label as found in the source text is in label. If this is None the resolver deleted the label definition.

val block : t -> block

block fn is the footnote content.

As label definitions

type Label.def +=

| Def of t node

A label definition for footnotes.