Module Block.List'

Lists.

type type' = [

| `Unordered of Layout.char
(*
with given marker.
*)

| `Ordered of int * Layout.char
(*
starting at given integer, markers ending with given character (')' or '.').
*)

]

The type for list types.

type t

The type for lists.

val make : ?tight:bool -> type' -> List_item.t node list -> t

make ?tight t items is a list with given parameters. tight default to true, but should be computed from items in practice.

val type' : t -> type'

type' l is the list type of l.

val tight : t -> bool

tight l is true iff the list is tight.

val items : t -> List_item.t node list

items l are the items of l.