Module Block.List_item

List items.

type block := t

type t

The type for list items.

val make : ?before_marker:Layout.indent -> ?marker:Layout.string node -> ?after_marker:Layout.indent -> ?ext_task_marker:Uchar.t node -> block -> t

make b is a list item for block b with given parameters, see corresponding accessors for semantics.

val block : t -> block

block i is the contents of i.

val before_marker : t -> Layout.indent

before_marker i is the indentation before the list marker.

val marker : t -> Layout.string node

marker i is the item marker layout of i.

val after_marker : t -> Layout.indent

after_marker i is the indentation after the marker.

Task items

val ext_task_marker : t -> Uchar.t node option

ext_task_marker i is a task marker, only occurs in non-strict parsing mode, see see extension list task items.

val task_status_of_task_marker : Uchar.t -> [ `Unchecked | `Checked | `Cancelled | `Other of Uchar.t ]

task_status_of_task_marker u is a status for marker u, see extension list task item.