Module Inline.Emphasis

Emphasis and strong emphasis.

type inline := t

type t

The type for emphasis and strong emphasis.

val make : ?delim:Layout.char -> inline -> t

make i is an emphasis on i. delim is the delimiter used it should be either '*' or '_'.

val inline : t -> inline

inline e is the emphasised inline.

val delim : t -> Layout.char

delim e is the delimiter used for emphasis, should be either '*' or '_'.