Module Inline.Math_span

Math span.

type t

The type for math spans.

val make : display:bool -> Block_line.tight list -> t

make tex_layout is an inline or display math span with given TEX code.

val display : t -> bool

display ms is true if the span should be on its own line.

val tex : t -> string

tex ms is the inline math TEX code of ms

val tex_layout : t -> Block_line.tight list

tex_layout ms is inline math TEX code in a form that allows layout preservation.

The acual code data is the tight block lines concatenated and separated by space. The tex function does that for you.