Module Cmarkit.Layout

Types for layout information.

Values of these types do not represent document data. They are used to recover document source layout informations when the abstract syntax tree cannot represent them. See source layout preservation for more information.

For programmatically generated nodes, values of these types can be left empty or filled with a desired layout. Except for the CommonMark renderer these values are usually ignored.

type blanks = string

The type for blanks layout. This is only made of spaces and tabs.

type nonrec string = string

The type for string layout. For example the art of thematic breaks or code fences.

type nonrec char = char

The type for character layout. For example the character used for an emphasis or an unordered list marker.

type count = int

The type for some kind of layout count. Usually a character count.

type indent = int

The type for block indentation. Mostly between 0-3.

val string : ?meta:Meta.t -> string -> string node

string s is a layout string with meta data meta (defaults to Meta.none).

val empty : string node

empty is string "".